Move classes related to FieldEditor to a separate gui package
