+++ src/main/java/net/sf/jabref/EntryEditor.java
@@ -49,6 +49,7 @@ import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
 import javax.swing.text.JTextComponent;
 
+import net.sf.jabref.gui.fieldeditors.*;
 import net.sf.jabref.logic.autocompleter.AutoCompleter;
 import net.sf.jabref.export.LatexFieldFormatter;
 import net.sf.jabref.external.ExternalFilePanel;
