+++ src/main/java/net/sf/jabref/EntryEditorTabFocusListener.java
@@ -15,6 +15,8 @@
 */
 package net.sf.jabref;
 
+import net.sf.jabref.gui.fieldeditors.FieldEditor;
+
 import javax.swing.*;
 import javax.swing.event.DocumentEvent;
 import javax.swing.event.DocumentListener;
