+++ src/main/java/net/sf/jabref/FieldContentSelector.java
@@ -32,6 +32,7 @@ import javax.swing.KeyStroke;
 
 import com.jgoodies.forms.layout.Sizes;
 import com.jgoodies.looks.Options;
+import net.sf.jabref.gui.fieldeditors.FieldEditor;
 import net.sf.jabref.util.Util;
 
 /**
