+++ src/main/java/net/sf/jabref/FieldTextMenu.java
@@ -46,6 +46,7 @@ import javax.swing.JPopupMenu;
 import javax.swing.KeyStroke;
 import javax.swing.text.JTextComponent;
 
+import net.sf.jabref.gui.fieldeditors.FieldEditor;
 import net.sf.jabref.util.CaseChangeMenu;
 import net.sf.jabref.logic.util.NameListNormalizer;
 import net.sf.jabref.util.GoogleUrlCleaner;
