+++ src/main/java/net/sf/jabref/PreambleEditor.java
@@ -20,6 +20,8 @@ import java.awt.*;
 
 import javax.swing.*;
 
+import net.sf.jabref.gui.fieldeditors.FieldEditor;
+import net.sf.jabref.gui.fieldeditors.FieldTextArea;
 import net.sf.jabref.gui.undo.UndoablePreambleChange;
 
 class PreambleEditor extends JDialog {
