+++ src/main/java/net/sf/jabref/PreviewPanel.java
@@ -35,7 +35,7 @@ import javax.swing.event.HyperlinkListener;
 import net.sf.jabref.export.layout.Layout;
 import net.sf.jabref.export.layout.LayoutHelper;
 import net.sf.jabref.export.ExportFormats;
-import net.sf.jabref.gui.PreviewPanelTransferHandler;
+import net.sf.jabref.gui.fieldeditors.PreviewPanelTransferHandler;
 import net.sf.jabref.util.Util;
 
 /**
