+++ src/main/java/net/sf/jabref/gui/fieldeditors/FieldEditor.java
@@ -13,7 +13,7 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
-package net.sf.jabref;
+package net.sf.jabref.gui.fieldeditors;
 
 import net.sf.jabref.gui.AutoCompleteListener;
