+++ src/main/java/net/sf/jabref/gui/fieldeditors/JTextAreaWithHighlighting.java
@@ -13,7 +13,11 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
-package net.sf.jabref;
+package net.sf.jabref.gui.fieldeditors;
+
+import net.sf.jabref.Globals;
+import net.sf.jabref.JabRefPreferences;
+import net.sf.jabref.SearchTextListener;
 
 import java.awt.event.ActionEvent;
 import java.util.ArrayList;
