+++ src/test/java/net/sf/jabref/SearchTextListenerTest.java
@@ -1,5 +1,6 @@
 package net.sf.jabref;
 
+import net.sf.jabref.gui.fieldeditors.FieldTextArea;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
